Run go mod tidy
Improve logging
Seperate file for grpc handler
